% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPL2025IDconvert.R
\name{GPL2025IDconvert}
\alias{GPL2025IDconvert}
\alias{GPL2025IDconverter}
\title{Convert Chip ID of the GPL2025 to GeneBank Accession and ENTREZID}
\usage{
GPL2025IDconverter(myID)
}
\arguments{
\item{myID}{A vector of chip id}
}
\value{
Return a vector or a datafram
}
\description{
\code{GPL2025IDconverter} Convert Chip ID of the GPL2015 to GeneBank Accession and ENTREZID
}
\examples{
convert_id <- GPL2025IDconverter('Os.11182.1.S1_at')
}
\author{
Xiang Li <ynaulx@gmail.com>
}
